fix(calc): R19-3 — a subrange bound was Rust's saturating `as`, not C's `(int)` cast
`AA[2,3e9]` on an 8-element AA: compiled C returns an EMPTY selection
(all-zero buffer); the port returned `AA[2..8]`. The opposite answer, from one
cast.

C narrows a subrange bound with a bare `(int)` (`aCalcPerform.c:1526,1530`;
`sCalcPerform.c:1876,1883`). On x86-64 that is `cvttsd2si`, which answers
`INT32_MIN` for anything outside the 32-bit range and for NaN/±Inf. `j =
INT32_MIN` then hits `if (j < 0) j += k` and stays far negative, so the range
inverts and the copy loop never runs. Rust's `as i64` SATURATES instead — 3e9
survives as 3e9, clamps to `arraySize`, and selects the whole tail.

`cast::c_int` (-> `types::c_cast::f64_to_i32`) already owns exactly this cast
for the rest of the engine. These sites bypassed it.

Audit — anchor: every f64 -> integer cast in `crates/epics-base-rs/src/calc/engine/`
(`rg 'as i32|as i64|as u32|as usize'`, then filtered to those whose SOURCE is a
calc stack double, which is what `c_int` owns).

  same defect, fixed:
    array.rs  `pop_subrange_bounds`  aCalc `[` / `{` — both bounds
    string.rs `subrange_bounds`      sCalc `[` — both DOUBLE bounds (the STRING
                                     bounds are a strstr, no cast involved)
    array.rs  `ArrayOp::NSmooth`     the pass count. C is `int j; j = ps->d;`
                                     (`:581`) — the same implicit `(int)` — and
                                     the loop runs `max(j,0)` times. Under `as
                                     usize`, `NSMOO(AA,1e30)` asked for
                                     `usize::MAX` passes and the engine never
                                     returned. It is now a no-op, as in C.

  distinct, skip:
    cast.rs internals              — this IS the owner (`d2i`/`d2ui`/`c_int`/
                                     `my_nint`, each a different C cast).
    string.rs `render_int`         — its `l` already came through `my_nint`,
                                     i.e. through the owner; the `as i32` there
                                     is C's `%d` reading an int out of the
                                     vararg, not a double conversion.
    cvt.rs `value as i64`          — guarded by the branch above it to
                                     |v| <= 1e7, so it cannot leave i32 range;
                                     it is C's `(epicsInt32)flt_value` on a
                                     value C has range-checked the same way.
    scanf.rs `v as i32/u32/...`    — `v` is a PARSED integer, and the narrowing
                                     is glibc `sscanf`'s store through the
                                     output object, not calc's `(int)` of a
                                     computed double. Different C function;
                                     its own defect if any (reported, unfixed).
    `*idx as usize`, `*nargs as usize`, `.len() as i64`, window arithmetic
                                   — casts of opcode operands and buffer
                                     lengths. No double, no C cast.

Ground truth: `aCalcPerform.c` + `aCalcPostfix.c` + `calcUtil.c` +
`myFreeListLib.c` compiled on this host against base's libCom, arraySize 8,
AA=[10..80]. Every expectation in the test file is that binary's output:

  AA[2,3]           [30,40,0,0,0,0,0,0]
  AA[2,3e9]         [0,0,0,0,0,0,0,0]
  AA[2,1e10]        [0,0,0,0,0,0,0,0]
  AA[-2,-1]         [70,80,0,0,0,0,0,0]
  AA[0/0,3]         [10,20,30,40,0,0,0,0]     (NaN -> INT32_MIN -> clamped to 0)
  AA[6,2147483647]  [70,80,0,0,0,0,0,0]       (the last value the cast holds)
  NSMOO(AA,1e30)    [10,20,30,40,50,60,70,80] (zero passes)
  NSMOO(AA,-1)      [10,20,30,40,50,60,70,80]

Tier 2 SEMANTICS, matching C: no deviation.

Tests: crates/epics-base-rs/tests/calc_subrange_index_cast.rs — one case per
boundary of the cast: in range, past INT32_MAX, at INT32_MAX, NaN, negative
(the wrap), and the same boundary through sCalc's string subrange and through
NSMOOTH's pass count.

+++ crates/epics-base-rs/src/calc/engine/string.rs
@@ -2073,8 +2073,10 @@ impl Extremum {
 fn subrange_bounds(subject: &[u8], start: &StackValue, end: &StackValue) -> (i64, i64) {
     let k = subject.len() as i64;
     let i = match start {
+        // C `i = (int)ps1->d` (`sCalcPerform.c:1876`) — [`c_int`], not Rust's
+        // saturating `as`, exactly as in aCalc's `[` (`pop_subrange_bounds`).
         StackValue::Double(d) => {
-            let i = *d as i64;
+            let i = i64::from(c_int(*d));
             if i < 0 { i + k } else { i }
         }
         StackValue::Str(needle) => {
@@ -2082,8 +2084,9 @@ fn subrange_bounds(subject: &[u8], start: &StackValue, end: &StackValue) -> (i64
         }
     };
     let j = match end {
+        // C `j = (int)ps2->d` (`sCalcPerform.c:1883`).
         StackValue::Double(d) => {
-            let j = *d as i64;
+            let j = i64::from(c_int(*d));
             if j < 0 { j + k } else { j }
         }
         StackValue::Str(needle) if needle.is_empty() => k,
